feat: support live TTS of fish audio
