+++ requirements.txt
@@ -22,7 +22,7 @@ edge-tts==6.1.10
 exceptiongroup==1.2.0; python_version < "3.11"
 frozenlist==1.4.1
 google-ai-generativelanguage==0.6.10
-google-api-core==2.15.0
+google-api-core[grpc]==2.15.0
 google-api-python-client==2.125.0
 google-auth==2.26.1
 google-auth-httplib2==0.2.0
@@ -37,7 +37,8 @@ grpcio-status==1.60.0
 h11==0.14.0
 httpcore==1.0.5
 httplib2==0.22.0
-httpx==0.27.2
+httpx-ws==0.6.2
+httpx[socks]==0.27.2
 idna==3.7
 jiter==0.5.0
 jsonpatch==1.33
@@ -83,13 +84,13 @@ socksio==1.0.0
 soupsieve==2.5
 sqlalchemy==2.0.25
 tenacity==8.2.3
-tetos==0.3.1
+tetos==0.4.0
 tqdm==4.66.1
 typing-extensions==4.12.2
 typing-inspect==0.9.0
 uritemplate==4.1.1
 urllib3==2.1.0
 websocket-client==1.8.0
-websockets==12.0
+wsproto==1.2.0
 yarl==1.14.0
 zhipuai==2.1.5.20230904
